backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
bcrypt==3.2.0
SQLAlchemy==1.4.23
PyMySQL==1.0.2 
python-pam==2.0.2 
psutil==5.9.0
Werkzeug==2.0.1 
mysql-connector-python==9.3.0
cryptography>=3.0.0
Jinja2>=2.11.0 